third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-content-sized-columns-resolution.html

<html>
<head>
<link href="resources/grid.css" rel="stylesheet"/>
<style>
 .grid {
     font: 10px/1 Ahem;
 }
 .gridMinContentFixedAndAuto {
     grid-template-columns: minmax(min-content, 15px) auto;
 }
 .gridMaxContentFixedAndAuto {
     grid-template-columns: minmax(max-content, 15px) auto;
 }
 .gridAutoAndAuto {
     grid-template-columns: auto auto;
 }
 .gridMinContentAndMinContentFixed {
     grid-template-columns: min-content minmax(min-content, 30px);
 }
 .gridMinContentAndMaxContentFixed {
     grid-template-columns: min-content minmax(max-content, 30px);
 }
 .gridMaxContentAndMinContent {
     grid-template-columns: max-content min-content;
 }
 .gridFixedMinContentAndMaxContent {
     grid-template-columns: minmax(10px, min-content) max-content;
 }
 .gridFixedMaxContentAndMinContent {
     grid-template-columns: minmax(10px, max-content) min-content;
 }
 .gridAutoMinContent {
     grid-template-columns: auto min-content;
 }
 .gridAutoMaxContent {
     grid-template-columns: auto max-content;
 }
 .gridMaxContentAndMinContentFixed {
     grid-template-columns: max-content minmax(min-content, 35px);
 }
 .gridMaxContentAndMaxContentFixed {
     grid-template-columns: max-content minmax(max-content, 35px);
 }

</style>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div class="grid gridMinContentFixedAndAuto" id="gridMinContentFixedAndAuto">
  <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridAutoAndAuto" id="gridAutoAndAuto">
  <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridMinContentAndMinContentFixed" id="gridMinContentAndMinContentFixed">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridMaxContentAndMinContent" id="gridMaxContentAndMinContent">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridFixedMinContentAndMaxContent" id="gridFixedMinContentAndMaxContent">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridFixedMaxContentAndMinContent" id="gridFixedMaxContentAndMinContent">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridMinContentAndMaxContentFixed" id="gridMinContentAndMaxContentFixed">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="constrainedContainer">
    <div class="grid gridMaxContentFixedAndAuto" id="gridMaxContentFixedAndAuto">
	<div class="firstRowBothColumn">XXXX XXXX</div>
    </div>
</div>

<div class="grid gridAutoMinContent" id="gridAutoMinContent">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="grid gridAutoMaxContent" id="gridAutoMaxContent">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<div class="constrainedContainer">
<div class="grid gridMaxContentAndMinContentFixed" id="gridMaxContentAndMinContentFixed">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>
</div>

<div class="grid gridMaxContentAndMaxContentFixed" id="gridMaxContentAndMaxContentFixed">
    <div class="firstRowBothColumn">XXXX XXXX</div>
</div>

<script>
   function testGridColumnsValues(id, computedColumnValue)
   {
       shouldBeEqualToString("window.getComputedStyle(" + id + ", '').getPropertyValue('grid-template-columns')", computedColumnValue);
   }

   testGridColumnsValues("gridMinContentFixedAndAuto", "15px 75px");
   testGridColumnsValues("gridAutoAndAuto", "45px 45px");
   testGridColumnsValues("gridMinContentAndMinContentFixed", "20px 30px");
   testGridColumnsValues("gridMaxContentAndMinContent", "70px 20px");
   testGridColumnsValues("gridFixedMinContentAndMaxContent", "10px 80px");
   testGridColumnsValues("gridFixedMaxContentAndMinContent", "60px 30px");
   testGridColumnsValues("gridMinContentAndMaxContentFixed", "20px 70px");
   testGridColumnsValues("gridMaxContentFixedAndAuto", "65px 25px");
   testGridColumnsValues("gridAutoMinContent", "70px 20px");
   testGridColumnsValues("gridAutoMaxContent", "20px 70px");
   testGridColumnsValues("gridMaxContentAndMinContentFixed", "70px 20px");
   testGridColumnsValues("gridMaxContentAndMaxContentFixed", "55px 35px");
</script>
</body>
</html>
